+++ angle.h
@@ -11,24 +11,24 @@ public:
 	friend class TestAngle;
 
 	Angle() : radians(0.00) {};
-	Angle(float radians) { this->radians = radians; };
+	Angle(double radians) { this->radians = radians; };
 
-	void setRadians(float radians);
-	void setDegrees(float degrees);
-	void setDxDy(float dx, float dy);
+	void setRadians(double radians);
+	void setDegrees(double degrees);
+	void setDxDy(double dx, double dy);
 	void setUp() { this->radians = (M_PI / 2); };
 	void setDown() { this->radians = ((3 * M_PI) / 2); };
 	void setLeft() { this->radians = M_PI; };
 	void setRight() { this->radians = 0; };
-	void rotate(float radians);
-	float getDegrees() const;
-	float getRadians() const;
-	float getDx() const { return 0.00; };
-	float getDy() const { return 0.00; };
+	void rotate(double radians);
+	double getDegrees() const;
+	double getRadians() const;
+	double getDx() const { return 0.00; };
+	double getDy() const { return 0.00; };
 private:
-	float radiansToDegrees() const;
+	double radiansToDegrees() const;
 	void normalizeRadians();
 
-	float radians;
+	double radians;
 };
